miRkwood: Do not download mirdup data if already present

+++ cgi-bin/install-data.sh
@@ -9,14 +9,19 @@ if [ -z "$1" ]
 	exit 1
 fi
 ROOT_PATH=$(readlink -f $1)
-cd $ROOT_PATH
 
-echo "Deploying Mirdup data"
-MIRDUP=$ROOT_PATH/mirdup
-mkdir $MIRDUP
-cd $MIRDUP
-wget http://www.cs.mcgill.ca/~blanchem/mirdup/model_plants.zip
-unzip model_plants.zip
-rm model_plants.zip
 cd $ROOT_PATH
+MIRDUP=$ROOT_PATH/mirdup
+if [ ! -f $MIRDUP/plant.model ]
+then
+    echo "Deploying Mirdup data"
+    mkdir $MIRDUP
+    cd $MIRDUP
+    wget http://www.cs.mcgill.ca/~blanchem/mirdup/model_plants.zip
+    unzip model_plants.zip
+    rm model_plants.zip
+    cd $ROOT_PATH
+else
+    echo "Mirdup data already there"
+fi
